worker/requirements.txt
pyyaml
beanstalkc
docker-py
requests
